rustc_trans: don't emit ZST allocas that are only assigned to.
